requirements.txt
streamlit>=1.32
plotly>=5.20
pandas>=2.1
requests>=2.31
python-dotenv>=1.0
